refactor(web): infer PNCPContract from zod schema
Drops the hand-maintained PNCPContract interface (plus unused
PNCPContractItem, anoCompra, sequencialCompra) and exports the type
inferred from PNCPContractSchema instead. The two parsePncp* helpers
lose their `as unknown as PNCPContract` casts, so API drift now has a
single source of truth.
